+++ CopilotKit/packages/sdk-js/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @copilotkit/sdk-js
 
+## 1.8.14-next.5
+
+### Patch Changes
+
+- @copilotkit/shared@1.8.14-next.5
+
+## 1.8.14-next.4
+
+### Patch Changes
+
+- @copilotkit/shared@1.8.14-next.4
+
+## 1.8.14-next.3
+
+### Patch Changes
+
+- @copilotkit/shared@1.8.14-next.3
+
+## 1.8.14-next.2
+
+### Patch Changes
+
+- @copilotkit/shared@1.8.14-next.2
+
 ## 1.8.14-next.1
 
 ### Patch Changes